refactored and added negative case e2e test
