+++ src/applications/check-in/tests/feature-flip/enabled.cypress.spec.js
@@ -1,4 +1,4 @@
-import features from './mocks/enabled.json';
+import features from '../mocks/enabled.json';
 
 describe('Check In Experience -- ', () => {
   beforeEach(() => {
