--- src/applications/check-in/tests/happy-path/mocks/enabled.json
@@ -1,11 +1,0 @@
-{
-  "data":{
-     "type":"feature_toggles",
-     "features":[
-        {
-           "name":"check_in_experience_enabled",
-           "value":true
-        }
-     ]
-  }
-}
